% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfun.R
\name{acsv}
\alias{acsv}
\title{reads the csv at filepath (no ".csv" needed)
assigns dataframe to object with the same name as the csv then returns the object
e.g.: acsv("random_data") is equivalent to:
random_data <- read.csv("random_data.csv")}
\usage{
acsv(filepath)
}
\description{
reads the csv at filepath (no ".csv" needed)
assigns dataframe to object with the same name as the csv then returns the object
e.g.: acsv("random_data") is equivalent to:
random_data <- read.csv("random_data.csv")
}
