% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfun.R
\name{cut_to_times}
\alias{cut_to_times}
\title{Cut to times}
\usage{
cut_to_times(data, data_time_column, times)
}
\arguments{
\item{data}{The main data set.}

\item{data_time_column}{The time column of main data set.}

\item{times}{The time column of the target data set}
}
\description{
Not recommended. Use \code{\link{merge_by_times}} instead.
}
\details{
Use when you have two data sets with date and time columns,
and you wish to combine the data sets by matching each data point in the
main set to the closest time point in the target set. This function will
return a data set identical to the main set, but with the target time column
and with fewer data points.
}
