% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfun.R
\name{acsv}
\alias{acsv}
\title{Assign CSV}
\usage{
acsv(filepath)
}
\description{
Read a CSV and assign it to a variable of the same name. It also returns
the data frame, but isn't meant to be used in a pipeline.
}
\examples{
# The following two lines of code are identical:
acsv("your_data")
your_data <- read.csv("your_data.csv")

}
