% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfun.R
\name{collapse_dt}
\alias{collapse_dt}
\title{combines date and time columns into a numerical column
dcol: date column; tcol: time column.
format: optional. Defaults to "\%m/\%d/\%y \%H:\%M:\%S"
rmdt: remove collapsed date and time columns? Defaults to true.}
\usage{
collapse_dt(data, dcol, tcol, format, rmdt)
}
\description{
combines date and time columns into a numerical column
dcol: date column; tcol: time column.
format: optional. Defaults to "\%m/\%d/\%y \%H:\%M:\%S"
rmdt: remove collapsed date and time columns? Defaults to true.
}
