% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfun.R
\name{collapse_dt}
\alias{collapse_dt}
\title{Collapse date and time}
\usage{
collapse_dt(data, dcol, tcol, format, rmdt)
}
\arguments{
\item{dcol}{Date column.}

\item{tcol}{Time column}

\item{format}{Optional. Defaults to "\%m/\%d/\%y \%H:\%M:\%S"}

\item{rmdt}{If true, removes old date and time columns. Defaults to true.}
}
\description{
Combines date and time columns in character format
into a single, numerical column using POSIXct.
}
