% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfun.R
\name{cut_to_times}
\alias{cut_to_times}
\title{selects timepoints from dataset that are closest in time to a shorter list
data: the dataset
data_time_column: the time column of the dataset to reference. No quotes needed.
times: list of times to cut down to
Note: "times" is a column, not the entire short dataset.
Note: the time column of the dataset will be replaced with the target times.}
\usage{
cut_to_times(data, data_time_column, times)
}
\description{
selects timepoints from dataset that are closest in time to a shorter list
data: the dataset
data_time_column: the time column of the dataset to reference. No quotes needed.
times: list of times to cut down to
Note: "times" is a column, not the entire short dataset.
Note: the time column of the dataset will be replaced with the target times.
}
