% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfun.R
\name{\%csv\%}
\alias{\%csv\%}
\title{shorter infix version of acsv() for use in pipelines
use as follows:
"random_data" \%csv\% (
function1() \%>\%
function2() )
is equivalent to:
random_data <- read.csv("random_data.csv") \%>\%
function1() \%>\%
function2()
NOTE: pipeline MUST be in parentheses for \%csv\% infix to work!}
\usage{
filepath \%csv\% pipe
}
\description{
shorter infix version of acsv() for use in pipelines
use as follows:
"random_data" \%csv\% (
function1() \%>\%
function2() )
is equivalent to:
random_data <- read.csv("random_data.csv") \%>\%
function1() \%>\%
function2()
NOTE: pipeline MUST be in parentheses for \%csv\% infix to work!
}
