% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfun.R
\name{merge_by_times}
\alias{merge_by_times}
\title{data: larger dataset
target_data: shorter dataset
tcol: date_time column to reference (no quotes needed.)
NOTE: the date_time column must have the same name for both datasets!
returned dataset will have the dates from target_data and columns from both datasets.}
\usage{
merge_by_times(data, target_data, tcol)
}
\description{
data: larger dataset
target_data: shorter dataset
tcol: date_time column to reference (no quotes needed.)
NOTE: the date_time column must have the same name for both datasets!
returned dataset will have the dates from target_data and columns from both datasets.
}
